% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\docType{data}
\name{benchmark.df}
\alias{benchmark.df}
\title{Knapsack Bruteforce Benchmark}
\format{A \code{data.frame} consisting of three vectors.
\describe{
  \item{noItems}{Number of items.}
  \item{single}{Computational time for single core in milliseconds.}
  \item{parallel}{Computational time for multi core in milliseconds.}
}}
\usage{
benchmark.df
}
\description{
A \code{data.frame} which holds data for 2..18 items using the knapsack bruteforce method with single and multi core.
}
\keyword{datasets}
