% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame with the colums v (value) and w (weight) as numerical values.}

\item{W}{Maximum weight.}

\item{selector}{A combination of TRUE/FALSE to specify which items to pick.}
}
\value{
Returns the best combination of objects to pick, so that v is max.

Returns the v (value), w (weight) and the s (selector) depending on the selector s.
}
\description{
brute_force_knapsack
}
