% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{combination_object_output}
\alias{combination_object_output}
\title{combination_object_output}
\usage{
combination_object_output(obj)
}
\arguments{
\item{obj}{A list having numerical entries for v (value), w (weight) and s (selector), a numerical vector with TRUE/FALSE entries}
}
\value{
Returns a list with the desired formatting for displaying the object.
}
\description{
combination_object_output
}
