% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Any numeric value greater than 0.}

\item{b}{Any numeric value greater than 0.}
}
\value{
The function returns the greatest common denominator of a and b.
}
\description{
Euclidean Algorithm
}
\examples{
euclidean (6, 22)
euclidean (100, 1000)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
