% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webService.R
\name{fetchByKpi}
\alias{fetchByKpi}
\title{Fetch By KPI}
\usage{
fetchByKpi(kpi, municipality, year = c())
}
\arguments{
\item{kpi}{Id of the KPIs as a list.}

\item{municipality}{Id of the municipality.}

\item{year}{The years to fetch as a vector}
}
\value{
Returns a data.frame containing the requested data
}
\description{
Fetch By KPI
}
