% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webService.R
\name{fetchByMunicipality}
\alias{fetchByMunicipality}
\title{Fetch By Municipality}
\usage{
fetchByMunicipality(municipality, year)
}
\arguments{
\item{municipality}{Id of the municipality.}

\item{year}{The year as integer.}
}
\value{
Returns a data.frame containing all KPIs belonging to the municipality in the given year.
}
\description{
Fetch By Municipality
}
