% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webService.R
\name{fetchKpis}
\alias{fetchKpis}
\title{Fetch KPIs}
\usage{
fetchKpis()
}
\value{
Returns a data.frame containing all KPIs with their id and description.
}
\description{
Fetch KPIs
}
\examples{
fetchKpis()
}
