% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webService.R
\name{fetchMunicipalities}
\alias{fetchMunicipalities}
\title{Fetch Municipalities}
\usage{
fetchMunicipalities()
}
\value{
Returns a data.frame containing all municipalities.
}
\description{
Fetch Municipalities
}
\examples{
fetchMunicipalities()
}
