% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame with the colums v (value) and w (weight) as numerical values.}

\item{W}{Maximum weight.}
}
\value{
Returns the best combination of objects to pick, so that v is max.
}
\description{
greedy knapsack
}
