% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{knapsack_input_validation}
\alias{knapsack_input_validation}
\title{knapsack_input_validation}
\usage{
knapsack_input_validation(df, W)
}
\arguments{
\item{df}{The dataframe input which is to be validated.}

\item{W}{The weight which is to be validated.}
}
\value{
Nothing. Stops if validation detected wrong user input.
}
\description{
knapsack_input_validation
}
