% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example Graph from Wikipedia}
\format{A \code{data.frame} consisting of three vectors.
\describe{
  \item{v1}{First node of the edge.}
  \item{v2}{Second node of the edge.}
  \item{w}{Distance of the edge.}
}}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
wiki_graph
}
\description{
A \code{data.frame} which can be used as an example for the dijkstra algorithm.
}
\keyword{datasets}
