% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data.frame with the colums v (value) and w (weight) as numerical values.}

\item{W}{Maximum weight.}

\item{parallel}{Set to TRUE if execution should be parallelized. Default is FALSE.}
}
\value{
Returns the best combination of objects to pick, so that v is max.
}
\description{
brute_force_knapsack
}
