% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A \code{data.frame} which consist of 3 vectors named \code{v1}, \code{v1} and \code{w} and all must have the same length as shown in the example. \code{v1} and \code{v2} represent one edge in the graph, while \code{w} holds the distance. Edges must be included twice, from \code{v1} to \code{v2} and the other way round.}

\item{init_node}{A scalar stating from which nodes the distances are to be calculated.}
}
\value{
Returns a vector consisting of the shortest paths from the node of origin \code{init_node} to the rest of the nodes.
}
\description{
Dijkstra Algorithm
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)

}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
