% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webService.R
\name{fetchByKpi}
\alias{fetchByKpi}
\title{Fetch By KPI}
\usage{
fetchByKpi(kpi, municipality, year = 0)
}
\arguments{
\item{kpi}{Id of the KPIs as a list. Seperator is a ','.}

\item{municipality}{Id of the municipality.}

\item{year}{(optional) The year as integer.}
}
\value{
Returns a data.frame containing the
}
\description{
Fetch By KPI
}
\examples{
fetchByKpi(list("N00914,U00405"), 1440, 2012)
}
