% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webService.R
\name{fetchByKpi}
\alias{fetchByKpi}
\title{Fetch By KPI}
\usage{
fetchByKpi(kpi, municipality, year = list())
}
\arguments{
\item{kpi}{Id of the KPIs as a list.}

\item{municipality}{Id of the municipality.}

\item{year}{The years to fetch as a list.}
}
\value{
Returns a data.frame containing the
}
\description{
Fetch By KPI
}
\examples{
fetchByKpi(list("N00914,U00405"), 1440, list(2010, 2011, 2012))
}
