% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Sample data for the Dijkstra's algorithm to find the shortest distances between the nodes/vertices}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
wiki_graph
}
\description{
Sample data for the Dijkstra's algorithm to find the shortest distances between the nodes/vertices
}
\keyword{datasets}
